@@ -1,8 +1,7 @@
-﻿namespace BadDicom.Configuration
+﻿namespace BadDicom.Configuration;
+
+internal class Config
 {
-    class Config
-    {
-        public TargetDatabase Database { get;set; }
-        public ExplicitUIDs UIDs { get; set; }
-    }
+    public TargetDatabase Database { get;set; }
+    public ExplicitUIDs UIDs { get; set; }
 }
