@@ -2,11 +2,11 @@ using System.Reflection;
 
 [assembly: AssemblyCompany("Health Informatics Centre, University of Dundee")]
 [assembly: AssemblyProduct("Bad Dicom")]
-[assembly: AssemblyCopyright("Copyright (c) 2018 - 2022")]
+[assembly: AssemblyCopyright("Copyright (c) 2018 - 2023")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
 // These should be replaced with correct values by the release process
-[assembly: AssemblyVersion("0.0.15")]
-[assembly: AssemblyFileVersion("0.0.15")]
-[assembly: AssemblyInformationalVersion("0.0.15")]
+[assembly: AssemblyVersion("0.0.16")]
+[assembly: AssemblyFileVersion("0.0.16")]
+[assembly: AssemblyInformationalVersion("0.0.16-rc1")]
